Split DatabaseInteractor
