--- src/main/java/org/hydev/mcpm/client/database/boundary/ListPackagesBoundary.java
@@ -1,17 +1,0 @@
-package org.hydev.mcpm.client.database.boundary;
-
-import org.hydev.mcpm.client.database.inputs.ListPackagesInput;
-import org.hydev.mcpm.client.database.results.ListPackagesResult;
-
-/**
- * Defines how users can interact with a boundary to receive paginated packages.
- */
-public interface ListPackagesBoundary {
-    /**
-     * Invoke this method to complete a list packages request.
-     *
-     * @param input See the ListPackagesInput record.
-     * @return See the ListPackagesResult record.
-     */
-    ListPackagesResult list(ListPackagesInput input);
-}
